requirements.txt
streamlit>=1.32
pandas>=2.1
PyYAML>=6.0
scikit-learn>=1.3
